refactor: Separate UserInfo updating from UserDataManager
