+++ Assets/Material/GUI/Images/frame_01_2.png.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 1632b73db42a0a8429f99743b0459506
+guid: ec308088638bb5f4fa307f518a4e7ac3
 TextureImporter:
   internalIDToNameTable: []
   externalObjects: {}
