--- Assets/Script/GameData/UserDataUpdate.cs
@@ -1,9 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class UserDataUpdate : MonoBehaviour
-{
-
-
-}
