--- Assets/Script/GameData/UserDataUpdate.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: cb42446fab09829488493c40adcfa90f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
