--- Assets/Script/Menu/UserInfo.cs
@@ -1,11 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-
-public static class UserInfo
-{
-    public static int Money;
-    public static bool[] IsUserItem = new bool[Constants.itemCount];
-    public static bool[] IsUserAchi = new bool[Constants.achiCount];
-}
